Sort age groups on the final union from antigens
